@@ -3,7 +3,7 @@ import { FetchData } from "./components/FetchData";
 import { Swagger } from "./components/Swagger";
 import CreateEventPage from "./Pages/Event/CreateEventPage";
 import CreateTeamPage from "./Pages/CreateTeam/CreateTeamPage";
-import { DonationPage } from "./Pages/Donation/DonationPage";
+import DonationPage  from "./Pages/Donation/DonationPage";
 import { TeamsListPage } from "./components/TeamsInfo/TeamsListPage";
 import { Home } from "./Pages/Home/Home";
 import LoginLanding from "./Pages/Login/LoginLanding";
@@ -11,6 +11,9 @@ import { TeamDetails } from "./components/TeamsInfo/TeamDetails";
 import { LoggedInUser } from "./components/TeamsInfo/LoggedInUser";
 import LoginButton from "./components/LoginButton";
 import SiteAdmin from "./Pages/SiteAdmin/SiteAdmin";
+import SuccessfulDonation from "./Pages/Donation/SuccessfulDonation";
+import FailedPaymentPage from "./Pages/Donation/FailedPaymentPage";
+import LinkRedirect from "./components/LinkRedirect/LinkRedirect";
 
 const AppRoutes = [
     {
@@ -47,7 +50,7 @@ const AppRoutes = [
          element: <TeamsListPage />
     },
     {
-        path:'/createEvent', 
+        path:`/createEvent`, 
         element: <CreateEventPage/>
     },
 
@@ -59,9 +62,7 @@ const AppRoutes = [
     {
         path: `/LoggedInUser`,
         element: <LoggedInUser />
-    },
-
-   
+    },   
     {
         path: `/LoginButton`,
         element: <LoginButton />
@@ -70,6 +71,23 @@ const AppRoutes = [
         path: `/AdminDashboard`,
         element: <SiteAdmin />
     },
+    {
+        path: `/successfuldonation/:personName/:teamName/:transactionId`, 
+        element:<SuccessfulDonation/>
+    },
+    {
+        path: `/failedpayment`,
+        element: <FailedPaymentPage />
+    },
+    {
+        path: `/links/:linkIdentifer`,
+        element: <LinkRedirect />
+    },
+    {
+        path: `/Donate/links/:linkIdentifer`,
+        element: <LinkRedirect />
+    }
+
 ];
 
 export default AppRoutes;
