@@ -50,171 +50,172 @@ const CreateEventForm = () => {
                 setCurrentEvent(defaultEvent);
             }
         }
-    };
+        };
 
-    const createEventHandler = async (event: React.FormEvent) => {
-        event.preventDefault()
+        const createEventHandler = async (event: React.FormEvent) => {
+            event.preventDefault()
 
-        var newEvent: Event = {
-            date: eventDate,
-            title: eventTitle,
-            location: eventLocation,
-            description: eventDescription,
-            mainImage: eventMainImage,
-            donationTarget: eventDonationTarget
+            var newEvent: Event = {
+                date: eventDate,
+                title: eventTitle,
+                location: eventLocation,
+                description: eventDescription,
+                mainImage: eventMainImage,
+                donationTarget: eventDonationTarget
+            }
+
+            try {
+                await EventsService.CreateEventViaAxios(newEvent)
+                nextCurrentEvent();
+                navigate("/")
+            } catch (e) {
+                alert("Could not add Event");
+            }
+
+
         }
 
-        try {
-            await EventsService.CreateEventViaAxios(newEvent)
-            nextCurrentEvent();
-            navigate("/")
-        } catch (e) {
-            alert("Could not add Event");
-        }
+        useEffect(() => {
+            if (Date.parse(eventDate.toString()) && eventTitle.trim().length !== 0 && eventLocation.trim().length !== 0 && eventDonationTarget > 0 && eventDescription.trim().length !== 0) {
+                setDisableSubmit(false)
+            }
+            else {
+
+                setDisableSubmit(true)
+            }
+        }, [eventTitle, eventLocation, eventDescription, eventDonationTarget, eventDate])
 
 
+        return (
+            <>
+                <h1 className="EventHeader">Create Event here</h1>
+                <div className="FormEventPageContentPosition">
+                    <Form onSubmit={createEventHandler} className="FormEventBorder">
+                        <FormGroup>
+
+                            <Row className="FormEventRowOne">
+                                <Col md={6} xs={8}>
+                                    <FormGroup>
+                                        <Label>
+                                            Event Title
+                                        </Label>
+
+                                        <Input
+                                            placeholder="Event Title"
+                                            aria-label="eventTitle"
+                                            data-testid="eventTitleInput"
+                                            onChange={e => setEventTitle(e.currentTarget.value)}
+                                        />
+                                    </FormGroup>
+                                </Col>
+
+                            </Row>
+                        </FormGroup>
+                        <FormGroup>
+
+                            <Row className="FormEventRowTwo">
+                                <Col md={6} xs={8}>
+                                    <FormGroup>
+                                        <Label>
+                                            Event Location
+                                        </Label>
+                                        <Input
+                                            placeholder="Event Location"
+                                            aria-label="eventLocation"
+                                            data-testid="eventLocationInput"
+
+                                            onChange={e => setEventLocation(e.currentTarget.value)}
+                                        />
+                                    </FormGroup>
+                                </Col>
+
+                            </Row>
+                        </FormGroup>
+                        <FormGroup>
+                            <Row className="FormEventRowThree">
+
+                                <Col md={6} xs={8}>
+
+
+                                    <Label>
+                                        Event Description
+                                    </Label>
+                                    <Input
+                                        type="textarea"
+                                        aria-label="eventDescription"
+                                        data-testid="eventDescriptionInput"
+
+                                        onChange={e => setEventDescription(e.currentTarget.value)}
+
+                                    />
+
+                                </Col>
+                            </Row>
+                        </FormGroup>
+                        <FormGroup>
+                            <Row className="FormEventRowFour">
+                                <Col md={6} xs={8}>
+
+                                    <Label>
+                                        Donation Target
+                                    </Label>
+                                    <Input
+                                        placeholder="$"
+                                        type="number"
+                                        aria-label="eventDonationGoal"
+                                        data-testid="eventDonationGoalInput"
+
+                                        onChange={e => setEventDonationTarget(Number(e.currentTarget.value))}
+
+                                    />
+                                </Col>
+                            </Row>
+                        </FormGroup>
+                        <FormGroup>
+                            <Row className="FormEventRowFive">
+                                <Col md={6} xs={8}>
+
+                                    <Label>
+                                        Event Start Date
+                                    </Label>
+                                    <Input
+                                        type="date"
+                                        aria-label="eventDate"
+                                        data-testid="eventDateInput"
+
+                                        onChange={e => setEventDate(e.currentTarget.valueAsDate!)}
+
+                                    />
+                                </Col>
+                            </Row>
+                        </FormGroup>
+                        <FormGroup>
+                            <Row className="FormEventRowSix">
+                                <Col md={6} xs={8}>
+                                    <Input
+                                        id="exampleFile"
+                                        name="file"
+                                        type="file"
+                                        placeholder="Team Logo"
+                                        onChange={e => setEventMainImage(e.target.value)}
+                                    />
+                                    <FormText>
+                                        Select an image that will be displayed as your team's logo
+                                    </FormText>
+                                </Col>
+                            </Row>
+                        </FormGroup>
+                        <Col md={12} xs={8} className="FormEventButtonPosition">
+
+                            <Button variant='contained' disabled={disableSubmit} className="FormEventButtonSubmit" type="submit" >Submit</Button>
+                        </Col>
+                    </Form>
+
+                </div>
+            </>
+
+        );
     }
 
-    useEffect(() => {
-        if (Date.parse(eventDate.toString()) && eventTitle.trim().length !== 0 && eventLocation.trim().length !== 0 && eventDonationTarget > 0 && eventDescription.trim().length !== 0) {
-            setDisableSubmit(false)
-        }
-        else {
-
-            setDisableSubmit(true)
-        }
-    }, [eventTitle, eventLocation, eventDescription, eventDonationTarget, eventDate])
-
-
-    return (
-        <>
-            <h1 className="EventHeader">Create Event here</h1>
-            <div className="FormEventPageContentPosition">
-                <Form onSubmit={createEventHandler} className="FormEventBorder">
-                    <FormGroup>
-
-                        <Row className="FormEventRowOne">
-                            <Col md={6} xs={8}>
-                                <FormGroup>
-                                    <Label>
-                                        Event Title
-                                    </Label>
-
-                                    <Input
-                                        placeholder="Event Title"
-                                        aria-label="eventTitle"
-                                        data-testid="eventTitleInput"
-                                        onChange={e => setEventTitle(e.currentTarget.value)}
-                                    />
-                                </FormGroup>
-                            </Col>
-
-                        </Row>
-                    </FormGroup>
-                    <FormGroup>
-
-                        <Row className="FormEventRowTwo">
-                            <Col md={6} xs={8}>
-                                <FormGroup>
-                                    <Label>
-                                        Event Location
-                                    </Label>
-                                    <Input
-                                        placeholder="Event Location"
-                                        aria-label="eventLocation"
-                                        data-testid="eventLocationInput"
-
-                                        onChange={e => setEventLocation(e.currentTarget.value)}
-                                    />
-                                </FormGroup>
-                            </Col>
-
-                        </Row>
-                    </FormGroup>
-                    <FormGroup>
-                        <Row className="FormEventRowThree">
-
-                            <Col md={6} xs={8}>
-
-
-                                <Label>
-                                    Event Description
-                                </Label>
-                                <Input
-                                    type="textarea"
-                                    aria-label="eventDescription"
-                                    data-testid="eventDescriptionInput"
-
-                                    onChange={e => setEventDescription(e.currentTarget.value)}
-
-                                />
-
-                            </Col>
-                        </Row>
-                    </FormGroup>
-                    <FormGroup>
-                        <Row className="FormEventRowFour">
-                            <Col md={6} xs={8}>
-
-                                <Label>
-                                    Donation Target
-                                </Label>
-                                <Input
-                                    placeholder="$"
-                                    type="number"
-                                    aria-label="eventDonationGoal"
-                                    data-testid="eventDonationGoalInput"
-
-                                    onChange={e => setEventDonationTarget(Number(e.currentTarget.value))}
-
-                                />
-                            </Col>
-                        </Row>
-                    </FormGroup>
-                    <FormGroup>
-                        <Row className="FormEventRowFive">
-                            <Col md={6} xs={8}>
-
-                                <Label>
-                                    Event Start Date
-                                </Label>
-                                <Input
-                                    type="date"
-                                    aria-label="eventDate"
-                                    data-testid="eventDateInput"
-
-                                    onChange={e => setEventDate(e.currentTarget.valueAsDate!)}
-
-                                />
-                            </Col>
-                        </Row>
-                    </FormGroup>
-                    <FormGroup>
-                        <Row className="FormEventRowSix">
-                            <Col md={6} xs={8}>
-                                <Input
-                                    id="exampleFile"
-                                    name="file"
-                                    type="file"
-                                    placeholder="Team Logo"
-                                    onChange={e => setEventMainImage(e.target.value)}
-                                />
-                                <FormText>
-                                    Select an image that will be displayed as your team's logo
-                                </FormText>
-                            </Col>
-                        </Row>
-                    </FormGroup>
-                    <Col md={12} xs={8} className="FormEventButtonPosition">
-
-                        <Button variant='contained' disabled={disableSubmit} className="FormEventButtonSubmit" type="submit" >Submit</Button>
-                    </Col>
-                </Form>
-
-            </div>
-        </>
-
-    );
-}
 
 export default CreateEventForm;
