@@ -17,7 +17,7 @@ const NavMenu = () => {
         { text: 'Home', href: '/' },
         { text: 'Add Event', href: '/createEvent' },
         { text: 'Admin Dashboard', href: '/AdminDashboard' },
-
+        { text: 'Donate', href: '/Donate' },
 
     ]
 
@@ -26,7 +26,7 @@ const NavMenu = () => {
     }
     
     const adminPages = (value: any) => {
-        if (isAdmin === false && value.key === "Add Event" && value.key === "Admin Dashboard") {
+        if ((isAdmin === false && value.key === "Add Event") || (isAdmin === false &&  value.key === "Admin Dashboard")) {
             return false
         }
         else {
