@@ -1,14 +1,32 @@
 import { Button, Card, Grid } from '@mui/material';
-import * as React from 'react';
+import  React, {useState, useEffect} from 'react';
 import { createSearchParams, useNavigate } from 'react-router-dom';
 import Team from '../../JsModels/team';
 import { authService } from '../../services/authService';
-
+import axios from 'axios'
 
 export const TeamCard = ({ id, name, description, mainImage, ownerID, eventID, donationTarget, isPublic, registrations }: Team) => {
     const navigate = useNavigate();
     //const user = authService.isLoggedIn();
     //{ localStorage.getItem("LoggedInUser") == "" ?}
+
+    const BaseUrl = process.env.PUBLIC_URL
+    const [loggedInUserId, setLoggedInUserId] = useState(0)
+
+    useEffect(() => {
+        const config = {
+            headers: { Authorization: `Bearer ${localStorage.getItem("access_token")}` }
+        };
+        const getUser = async () => {
+            await axios.get(BaseUrl + '/api/user', config).then((response) => {
+                setLoggedInUserId(response?.data?.id)
+            }).catch((error)=> {
+            })
+        }
+
+         getUser()
+    })
+
     const loggedInUSer = localStorage.getItem("LoggedInUser")
         return (
             <div style={{ paddingTop: "1rem", justifyContent: "flex-start" }}>
@@ -25,7 +43,7 @@ export const TeamCard = ({ id, name, description, mainImage, ownerID, eventID, d
                                             navigate({
                                                 pathname: '/TeamDetails',
                                                 search: `?${createSearchParams({
-                                                    id: `${id}`,
+                                                    teamId: `${id}`,
                                                     ownerID: `${ownerID}`
 
                                                 })}`
@@ -42,8 +60,8 @@ export const TeamCard = ({ id, name, description, mainImage, ownerID, eventID, d
                                                             navigate({
                                                                 pathname: "/LoggedInUser",
                                                                 search: `?${createSearchParams({
-                                                                    id: `${id}`,
-                                                                    ownerID: `${ownerID}`
+                                                                    teamId: `${id}`,
+                                                                    userId: `${loggedInUserId}`
 
                                                                 })}`
                                                             })
@@ -73,3 +91,5 @@ export const TeamCard = ({ id, name, description, mainImage, ownerID, eventID, d
 
         );
     }
+
+
