@@ -27,10 +27,9 @@
                 if (System.Diagnostics.Debugger.IsAttached)
                 {
                     logger.LogWarning("Maybe it's a connection string issue, or the database is not up?\n");
-                    logger.LogWarning(@"If so, try these commands FROM THE src/combined DIRECTORY:
+                    logger.LogWarning(@"If so, try this command FROM THE dev-resources/api DIRECTORY:
 
-dotnet user-secrets set ""ASPEN_CONNECTION_STRING"" ""server=localhost; port=5434; database=postgres; user id=postgres; password=P@assword1""
-docker run -d --name pg -e POSTGRES_USER=postgres -e POSTGRES_PASSWORD=P@assword1 -p 5434:5432 postgres
+docker compose up 
 
 ");
                 }
