@@ -1,4 +1,7 @@
-﻿namespace Api.Models.Entities;
+﻿using System.ComponentModel.DataAnnotations.Schema;
+using System.Transactions;
+
+namespace Api.Models.Entities;
 
 public record Donation
 {
@@ -12,4 +15,9 @@ public record Donation
     public DateTime Date { get; init; }
     public decimal Amount { get; init; }
     public bool IsPending { get; init; } = true;
+    public long? Link { get; init; }
+    public Guid TransactionNumber { get; init; }
+
+    public string? AuthorizationNumber { get; init; }
+    public string? LinkGuid { get; init; }
 }
